HV-1478 Make the tests more consistent with the rest of the project
