--- osgi/felixtest/cdi-validation/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Hibernate Validator, declare and validate application constraints
-  ~
-  ~ License: Apache License, Version 2.0
-  ~ See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
-  -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"> <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>org.hibernate.validator</groupId>
-        <artifactId>hibernate-validator-osgi-felixtest</artifactId>
-        <version>6.0.3-SNAPSHOT</version>
-    </parent>
-    
-    <artifactId>cdi-validation</artifactId>
-    <packaging>war</packaging>
-    
-    <name>Hibernate Validator OSGi integration test for Felix - CDI Validation</name>
-    
-    <properties>
-        <hibernate-validator-parent.path>../../..</hibernate-validator-parent.path>
-    </properties>
-    
-</project>
