+++ osgi/felixtest/src/test/java/com/example/cdi/MyBean.java
@@ -10,10 +10,10 @@ import javax.enterprise.context.RequestScoped;
 
 @RequestScoped
 public class MyBean {
-    
-    public void doDefault(@ValidNumber String number) {
-    }
 
-    public void doAll9(@ValidNumber("all9") String number) {
-    }
+	public void doDefault(@ValidNumber String number) {
+	}
+
+	public void doAll9(@ValidNumber("all9") String number) {
+	}
 }
