+++ osgi/felixtest/src/test/java/com/example/cdi/ValidNumberValidator.java
@@ -11,19 +11,19 @@ import javax.validation.ConstraintValidatorContext;
 
 public class ValidNumberValidator implements ConstraintValidator<ValidNumber, String> {
 
-    private ValidNumber validNumber;
+	private ValidNumber validNumber;
 
-    @Override
-    public void initialize(ValidNumber validNumber) {
-       this.validNumber = validNumber;
-    }
+	@Override
+	public void initialize(ValidNumber validNumber) {
+		this.validNumber = validNumber;
+	}
 
-    @Override
-    public boolean isValid(String value, ConstraintValidatorContext context) {
-        if ("all9".equals(validNumber.value())) {
-            return value.chars().allMatch(c -> c == '9');
-        }
-        
-        return true;
-    }
+	@Override
+	public boolean isValid(String value, ConstraintValidatorContext context) {
+		if ( "all9".equals( validNumber.value() ) ) {
+			return value.chars().allMatch( c -> c == '9' );
+		}
+
+		return true;
+	}
 }
